oi-405 ready for code review
added tableview to display differences between actual and nominal geometry (x, y, z). If the difference is out of the specified tolerance the text will red. You can copy the differences with cmd+ c or with the contextmenu. If you use cmd c the mouse cursor has to be over the tableview